third_party/blink/web_tests/platform/generic/fast/forms/time-multiple-fields/time-multiple-fields-fallback-format-expected.txt

Checks incomplete datetime formats are rejected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Valid format
PASS getUserAgentShadowTextContent(input) is "01.234:59:11 PM"
PASS getUserAgentShadowTextContent(input) is "01.234:59:23"
PASS getUserAgentShadowTextContent(input) is "01.234:59:23 PM"
Invalid format, fallback to HH:mm:ss
PASS getUserAgentShadowTextContent(input) is "23:59:01.234"
PASS getUserAgentShadowTextContent(input) is "23:59:01.234"
PASS getUserAgentShadowTextContent(input) is "23:59:01.234"
PASS successfullyParsed is true

TEST COMPLETE
